// src/demo/demo-schema.ts
/* tslint:disable */
/* eslint-disable */

/**
 * AUTO-GENERATED FILE - DO NOT EDIT!
 *
 * This file was automatically generated by pg-to-ts v.4.0.0
 * $ pg-to-ts generate -c postgres://username:password@localhost:5432/postgres -t comment -t doc -t users -s public
 *
 */

export type Json = unknown;
export type comment_status_enum =
  | 'archived'
  | 'complete'
  | 'deleted'
  | 'pending';

// Table comment
/** Variant-level comments */
export interface Comment {
  id: string;
  doc_id: string;
  author_id: string;
  created_at: Date | null;
  modified_at: Date | null;
  /** Additional comment info @type {CommentMetadata} */
  metadata: Json | null;
  /** Content of the comment, formatted with Markdown. May contain @mentions. */
  content_md: string;
  /** List of statuses; Just an array for testing! */
  statuses: comment_status_enum[] | null;
}
/** Variant-level comments */
export interface CommentInput {
  id?: string;
  doc_id: string;
  author_id: string;
  created_at?: Date | null;
  modified_at?: Date | null;
  /** Additional comment info @type {CommentMetadata} */
  metadata?: Json | null;
  /** Content of the comment, formatted with Markdown. May contain @mentions. */
  content_md: string;
  /** List of statuses; Just an array for testing! */
  statuses?: comment_status_enum[] | null;
}
const comment = {
  tableName: 'comment',
  columns: [
    'id',
    'doc_id',
    'author_id',
    'created_at',
    'modified_at',
    'metadata',
    'content_md',
    'statuses',
  ],
  requiredForInsert: ['doc_id', 'author_id', 'content_md'],
  primaryKey: 'id',
  foreignKeys: {
    doc_id: {table: 'doc', column: 'id', $type: null as unknown as Doc},
    author_id: {table: 'users', column: 'id', $type: null as unknown as Users},
  },
  $type: null as unknown as Comment,
  $input: null as unknown as CommentInput,
} as const;

// Table doc
export interface Doc {
  id: string;
  created_by: string;
  title: string | null;
  contents: string | null;
}
export interface DocInput {
  id?: string;
  created_by: string;
  title?: string | null;
  contents?: string | null;
}
const doc = {
  tableName: 'doc',
  columns: ['id', 'created_by', 'title', 'contents'],
  requiredForInsert: ['created_by'],
  primaryKey: 'id',
  foreignKeys: {
    created_by: {table: 'users', column: 'id', $type: null as unknown as Users},
  },
  $type: null as unknown as Doc,
  $input: null as unknown as DocInput,
} as const;

// Table users
export interface Users {
  id: string;
  name: string;
  pronoun: string | null;
}
export interface UsersInput {
  id?: string;
  name: string;
  pronoun?: string | null;
}
const users = {
  tableName: 'users',
  columns: ['id', 'name', 'pronoun'],
  requiredForInsert: ['name'],
  primaryKey: 'id',
  foreignKeys: {},
  $type: null as unknown as Users,
  $input: null as unknown as UsersInput,
} as const;

export interface TableTypes {
  comment: {
    select: Comment;
    input: CommentInput;
  };
  doc: {
    select: Doc;
    input: DocInput;
  };
  users: {
    select: Users;
    input: UsersInput;
  };
}

export const tables = {
  comment,
  doc,
  users,
};